fix: It's not possible to publish SQL server db locally in the same way as SQLite
Deleted the folder because each user has to publish the SQL server db locally on their own machine. Although you could publish it to a server, but I couldn't be bothered doing that. I'll do it in another future project. Also fyi this is not really a fix, but you propably already realized it by now.